+++ packages/components/src/popover/index.ts
@@ -1,4 +1,0 @@
-import { withInstall } from '../utils/install'
-import Popover from './popover.vue'
-
-export const EnPopover = withInstall(Popover)