+++ packages/components/src/popper/index.ts
@@ -1,0 +1,4 @@
+import { withInstall } from '../utils/install'
+import Popper from './popper.vue'
+
+export const EnPopper = withInstall(Popper)