+++ packages/components/src/utils/mock.ts
@@ -1,0 +1,4 @@
+/**
+ * mock random number in range [0, 1000]
+ */
+export const id = () => Math.floor(Math.random() * 10000)